adds mediatr, DI, updates structure, tests and mocks

+++ TripsApp.Api/TripsApp.Domain/Mocks/TripsMock.cs
@@ -11,21 +11,21 @@ namespace TripsApp.Domain.Mocks
                 new Trip
                 {
                     CountryId = 144,
-                    Distance = (decimal)55.6f,
+                    Distance = 55.6,
                     TimeStamp = new DateTime(2020,06,05),
                     VehicleId = new Guid()
                 },
                 new Trip
                 {
                     CountryId = 145,
-                    Distance = (decimal)155.6f,
+                    Distance = 155.6,
                     TimeStamp = new DateTime(2020,06,05),
                     VehicleId = new Guid()
                 },
                 new Trip
                 {
                     CountryId = 146,
-                    Distance = (decimal)160.00f,
+                    Distance = 160.00,
                     TimeStamp = new DateTime(2020,06,05),
                     VehicleId = new Guid()
                 }
